Implemented Bomb Weapon
- Created a Bomb Weapon (Grenade Launcher) and the associated scripts
(Bomb, BombWeapon, BombWeaponData)
- Created a Grenade Launcher with the following properties:
- Limited use (only 5 as of now; value can change) AOE weapon with high
knockback.
- Blast only hits players once.
- When exhausted, it will be replaced with the weapon the player picked
up.

--- POSSIBLE ISSUE?
- If a player holds onto an limited use weapon into the next round, they
will not have a weapon when that weapon is exhausted, unless that weapon
equips a new default weapon for the player.

--- POSSIBLE GAMEPLAY EXPLOIT
- If a player has carried a limited use weapon into the start of a new
round, and proceeds to not pick up a weapon that round, they will start
the following round with that weapon again. This allows a player to hold
onto a weapon for as long as they wish.