players can now see how much "ammo" they have.
The aiming arrow changes color depending on the type of weapon. If the
weapon type is Instant, they get a cyan colored arrow. If charge, its
white.

The Bomb and the Black Hole weapon have been modified to return a
ChargeLevel that is equal to the amount of "Ammo" they have left. This
required changing ChargeLevel in the Weapon class to check the type of
weapon. This is a hacky method of doing it, but whatever, it works.