grenade launcher projectile now bounces.
the projectine from the grenade launcher now bounces two times whenever
a player shoots it. Both explosions deal the same amount of knockback
and damage.

To change how the bouncing works, two variables need to be manipulated:
- the Arc Height variable in BombProjectileAddon on the Bomb prefab.
- the Projectile speed of the Bomb in Grenade Launcher.

Arc Height is *not* an accurate measure of how high it will go, because
gravity exists which pulls the bomb down. The actual height will be
roughly half of the Arc Height. ProjSpeed seems to make the arc higher
or lower, depending on how high the value is.

Two functions were added to the projectile class to get a boolean value
to make the bouncing possible. It shouldn't affect any of the existing
weapons at all.