Weapon changes + Imported Player Model
The grenade launcher now travels less distance than before, making it
more useful in the endgame.

The Black Hole weapon finally replaces itself after it  is used at the
start of a new round with a slingshot.

Imported an untextured player model that I made during Global Game Jam
into the project. To be removed if we make a better one.