Dodge : The Tweaks Go On
There is basically no difference till from 100+fps to 15fps. But once you dip below 15 the dodge goes a bit longer

Making it dodge based would take too long for me to do right now

The dodge currently decelerates a bit while dodging to make it look a bit smoother. But like we can set a value  ("DodgeDeceleration" in the data) to 0 to end that.