Bomb/Black Hole no longer carry over rounds.
The Bomb and Black Hole weapons no longer carry over into the following
rounds. Instead, the following will occur:

- If the player's current weapon is a Temp-type Weapon, they will be
made to use their previously equipped weapon at the start of the next
round.
- If both the player's current and previous weapons are Temp-type
Weapons, they will be made to use the default weapon.
- If the player's previous weapon is a Temp-type weapon, they will start
the next round with their current Weapon as long as it isn't a Temp-type
weapon.

I found that I could simply add some lines of code to the
PlayerController class to simply check the weapons it had when it gets
destroyed, and to fix it accordingly.