Added new map: Arena, Modified Slingshot
The simple Arena map has been made! It is a 27x27 map that has no real
gimmick to it. just 4 pillars, 4 players, a handful of items, and their
skill.

The slingshot has had some changes made to it to make it work better.
- It will always check its velocity to make sure it is moving at the
right speed.
- The hit detection for the ricochet now has a layer mask, ensuring it
will only interact with the Player layer and the Ground layer.