Added 3 new map alts; Reimplemented level selection
There are 3 new map alts that have been added to the game:
IslandsBridge, Arena Raised, and PitSquare. Each map belongs to a
specific level group to increase map variety.

The old method has been reimplemented, with a slight variation. It works
as follows:
- The ranks from 0 to the highest rank represent a range of values.
- This range is divided up into 4 groups of ranks. each of these groups
corresponds with one of 4 level groups.
- We check the rank of each player and see how many fall into each rank
group.
- Whichever rank group has the most amount of players determines which
level group we choose - i.e. if rank group 2 has the most players, we
pick level group 2.
- If there is a tie for which rank groups has the most players, we pick
the higher order rank group each time - i.e. if group 1 and 2 have an
equal number of players, we pick level group 2.
- Once we select a level group, we select a level from it. If the same
level group is chosen twice in a row, we pick the next level in the
group.

NOTE: The new level selection does not prevent itself from picking a
level group that is a lower order than the previously chosen level
group.